Add count query capability to MongoDB provider
A new function has been implemented in the MongoDB provider enabling it to count the number of documents in a collection matching a given query and return the count. This feature will generate a 'count' field in the response body when a count request is made. Additionally, modified the test suite structure for more effective testing.