Add date template functionality to MongoProvider
In the MongoProvider, added the ability to use date templates within queries, implementing a DayTemplate struct that formats specific dates within a week for use in such queries. Updated the README with information on the new functionality. Extended tests to cover this new feature. This commit aids in making queries more dynamic when date constraints are required.